APP-5472 Address encoding issue buy moving of upping of first and surnames to C# code
