+++ INSS.EIIR.DataAccess/DatabaseScripts/getEiirIndexBIVAonly.sql
@@ -193,8 +193,8 @@ FROM #Cases c
     individual.indiv_no AS indivNo, 
 
     -- Individual details        
-    ISNULL(NULLIF(UPPER(individual.forenames),''), 'No Forenames Found') AS individualForenames, 
-    ISNULL(NULLIF(UPPER(individual.surname), ''), 'No Surname Found') AS individualSurname, 
+    ISNULL(NULLIF(individual.forenames,''), 'No Forenames Found') AS individualForenames, 
+    ISNULL(NULLIF(individual.surname, ''), 'No Surname Found') AS individualSurname, 
     ISNULL(NULLIF(individual.title, ''), 'No Title Found') AS individualTitle,
 
     CASE
