+++ INSS.EIIR.DataSync.Application/UseCase/SyncData/AutoMapperProfiles/InsolventIndividualRegisterModelMapper.cs
@@ -7,6 +7,8 @@ using System.Threading.Tasks;
 
 using INSS.EIIR.DataSync.Application.UseCase.SyncData.Model;
 using INSS.EIIR.Models.CaseModels;
+using INSS.EIIR.Models;
+using Azure.Search.Documents.Indexes.Models;
 
 namespace INSS.EIIR.DataSync.Application.UseCase.SyncData.AutoMapperProfiles
 {
@@ -16,6 +18,9 @@ namespace INSS.EIIR.DataSync.Application.UseCase.SyncData.AutoMapperProfiles
         public InsolventIndividualRegisterModelMapper()
         {
             CreateMap<CaseResult, InsolventIndividualRegisterModel>()
+                //APP-5472 fix for 8 bit encoding issue
+                .ForMember(dest => dest.individualForenames, opt => opt.MapFrom(src => IirEncodingHelper.FixSQLEncoding(src.individualForenames)))
+                .ForMember(dest => dest.individualSurname, opt => opt.MapFrom(src => IirEncodingHelper.FixSQLEncoding(src.individualSurname)))
                 //CaseResult historically did not contain annulDate and annulReason derived from caseStatus
                 .ForMember(dest => dest.annulDate, opt => opt.MapFrom(src => src.caseStatus != null && src.caseStatus.StartsWith("ANNULLED", StringComparison.OrdinalIgnoreCase) ? src.caseStatus.TakeLast(10) : null))
                 .ForMember(dest => dest.annulReason, opt => opt.MapFrom(src => src.caseStatus != null && src.caseStatus.StartsWith("ANNULLED", StringComparison.OrdinalIgnoreCase) ? src.caseStatus.Take(src.caseStatus.Length-14) : null))
